Add error notify if JACK fails to start

+++ scripts/audio_setup.sh
@@ -37,11 +37,23 @@ else
 fi
 
 jack_control start | tee -a $LOG
+
+if jack_control status | grep "stopped"
+then
+    notify-send -u critical "JACK Failed to Start"
+    exit
+fi
+
 jack_control dps rate 48000 | tee -a $LOG
 jack_control eps realtime true
 jack_control dps nperiods 3 | tee -a $LOG
 jack_control dps period 256 | tee -a $LOG
 sleep 10
+
+if [jack_control status | grep "stopped" ]; then
+    notify-send -u "JACK Failed to Start"
+    exit
+fi
 
 echo -e "Connecting a2j MIDI \n" >> $LOG
 killall -9 a2jmidid
